requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
config==0.5.1
idna==3.4
lxml==4.9.2
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
